move info-icon to top in order to reuse
